+++ app/Actions/Import/FromUrl.php
@@ -6,8 +6,8 @@ use App\Actions\Photo\Create;
 use App\Actions\Photo\Strategies\ImportMode;
 use App\Exceptions\Handler;
 use App\Exceptions\MassImportException;
-use App\Image\BaseMediaFile;
-use App\Image\DownloadedFile;
+use App\Image\Files\BaseMediaFile;
+use App\Image\Files\DownloadedFile;
 use App\Models\Album;
 use App\Models\Configs;
 use App\Models\Photo;
