+++ app/Actions/Photo/Archive.php
@@ -8,7 +8,7 @@ use App\Contracts\Models\AbstractSizeVariantNamingStrategy;
 use App\Exceptions\ConfigurationKeyMissingException;
 use App\Exceptions\Internal\FrameworkException;
 use App\Exceptions\Internal\InvalidSizeVariantException;
-use App\Image\FlysystemFile;
+use App\Image\Files\FlysystemFile;
 use App\Models\Configs;
 use App\Models\Photo;
 use App\Models\SizeVariant;
