+++ app/Actions/Photo/Create.php
@@ -18,8 +18,8 @@ use App\Exceptions\Internal\LycheeAssertionError;
 use App\Exceptions\Internal\QueryBuilderException;
 use App\Exceptions\InvalidPropertyException;
 use App\Exceptions\MediaFileOperationException;
-use App\Image\BaseMediaFile;
-use App\Image\NativeLocalFile;
+use App\Image\Files\BaseMediaFile;
+use App\Image\Files\NativeLocalFile;
 use App\Image\StreamStat;
 use App\Metadata\Extractor;
 use App\Models\Album;
