+++ app/Actions/Photo/Strategies/RotateStrategy.php
@@ -13,8 +13,8 @@ use App\Exceptions\Internal\InvalidRotationDirectionException;
 use App\Exceptions\Internal\LycheeAssertionError;
 use App\Exceptions\Internal\LycheeDomainException;
 use App\Exceptions\MediaFileUnsupportedException;
-use App\Image\FlysystemFile;
-use App\Image\ImageHandler;
+use App\Image\Files\FlysystemFile;
+use App\Image\Handlers\ImageHandler;
 use App\Models\Photo;
 use App\Models\SizeVariant;
 use Illuminate\Contracts\Container\BindingResolutionException;
