+++ app/Console/Commands/ExifLens.php
@@ -5,7 +5,7 @@ namespace App\Console\Commands;
 use App\Contracts\Exceptions\ExternalLycheeException;
 use App\Exceptions\ModelDBException;
 use App\Exceptions\UnexpectedException;
-use App\Image\BaseMediaFile;
+use App\Image\Files\BaseMediaFile;
 use App\Metadata\Extractor;
 use App\Models\Photo;
 use App\Models\SizeVariant;
