+++ app/Contracts/Image/ImageHandlerInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Image;
+namespace App\Contracts\Image;
 
 use App\DTO\ImageDimension;
 use App\Exceptions\ImageProcessingException;
@@ -16,7 +16,7 @@ interface ImageHandlerInterface
 	/**
 	 * Loads an image from the provided file.
 	 *
-	 * @param BaseMediaFile $file the file to read from
+	 * @param MediaFile $file the file to read from
 	 *
 	 * @return void
 	 *
@@ -24,19 +24,19 @@ interface ImageHandlerInterface
 	 * @throws MediaFileOperationException
 	 * @throws ImageProcessingException
 	 */
-	public function load(BaseMediaFile $file): void;
+	public function load(MediaFile $file): void;
 
 	/**
 	 * Save the image into the provided file.
 	 *
-	 * @param BaseMediaFile $file              the file to write into
+	 * @param MediaFile $file              the file to write into
 	 * @param bool          $collectStatistics if true, the method returns statistics about the stream
 	 *
-	 * @return StreamStat|null optional statistics about the stream, if requested
+	 * @return StreamStats|null optional statistics about the stream, if requested
 	 *
 	 * @throws MediaFileOperationException
 	 */
-	public function save(BaseMediaFile $file, bool $collectStatistics = false): ?StreamStat;
+	public function save(MediaFile $file, bool $collectStatistics = false): ?StreamStats;
 
 	/**
 	 * Frees all internal resources.
