+++ app/Http/Controllers/PhotoController.php
@@ -26,8 +26,8 @@ use App\Http\Requests\Photo\SetPhotosStarredRequest;
 use App\Http\Requests\Photo\SetPhotosTagsRequest;
 use App\Http\Requests\Photo\SetPhotosTitleRequest;
 use App\Http\Requests\Photo\SetPhotoUploadDate;
-use App\Image\TemporaryLocalFile;
-use App\Image\UploadedFile;
+use App\Image\Files\TemporaryLocalFile;
+use App\Image\Files\UploadedFile;
 use App\ModelFunctions\SymLinkFunctions;
 use App\Models\Configs;
 use App\Models\Photo;
