+++ app/Image/Files/InMemoryBuffer.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace App\Image;
+namespace App\Image\Files;
 
 use App\Exceptions\MediaFileOperationException;
+use App\Image\StreamStat;
 
 class InMemoryBuffer extends AbstractBinaryBlob
 {
