+++ app/Image/Files/NativeLocalFile.php
@@ -1,9 +1,11 @@
 <?php
 
-namespace App\Image;
+namespace App\Image\Files;
 
 use App\Exceptions\MediaFileOperationException;
 use App\Exceptions\MediaFileUnsupportedException;
+use App\Image\StreamStat;
+
 use function Safe\filemtime;
 use function Safe\filesize;
 use function Safe\fopen;
