+++ app/Image/SizeVariantDefaultFactory.php
@@ -3,6 +3,7 @@
 namespace App\Image;
 
 use App\Contracts\Exceptions\LycheeException;
+use App\Contracts\Image\ImageHandlerInterface;
 use App\Contracts\Models\AbstractSizeVariantNamingStrategy;
 use App\Contracts\Models\SizeVariantFactory;
 use App\DTO\ImageDimension;
@@ -14,6 +15,9 @@ use App\Exceptions\Internal\IllegalOrderOfOperationException;
 use App\Exceptions\Internal\InvalidSizeVariantException;
 use App\Exceptions\MediaFileOperationException;
 use App\Exceptions\MediaFileUnsupportedException;
+use App\Image\Files\TemporaryLocalFile;
+use App\Image\Handlers\ImageHandler;
+use App\Image\Handlers\VideoHandler;
 use App\Models\Configs;
 use App\Models\Photo;
 use App\Models\SizeVariant;
